wk-2/week2/src/main/java/frank/CommonIntegrationTestingStrategies/TopDownApproach/LibraryManagementSystem/src/main/java/frank/librarymanagementsystem/controllers/BookController.java
package frank.librarymanagementsystem.controllers;

import frank.librarymanagementsystem.models.Book;
import frank.librarymanagementsystem.services.BookService;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/books")
@RequiredArgsConstructor
public class BookController {

    private final BookService bookService;

    @GetMapping("/search")
    public List<Book> searchBooks(@RequestParam String keyword) {
        return bookService.searchBooks(keyword);
    }
}